update: 선물 내역 이메일 전송 기능 비동기 처리 (#65)
