+++ src/main/java/com/sharetreats/chatbot/module/service/PaymentServiceImpl.java
@@ -6,10 +6,12 @@ import com.sharetreats.chatbot.module.dto.GiftHistoryRedisHash;
 import com.sharetreats.chatbot.module.entity.Account;
 import com.sharetreats.chatbot.module.entity.GiftHistory;
 import com.sharetreats.chatbot.module.entity.Product;
+import com.sharetreats.chatbot.module.event.GiftHistoryEmailEvent;
 import com.sharetreats.chatbot.module.repository.AccountRepository;
 import com.sharetreats.chatbot.module.repository.GiftHistoryRepository;
 import com.sharetreats.chatbot.module.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -28,8 +30,7 @@ public class PaymentServiceImpl implements PaymentService {
     private final ProductRepository productRepository;
     private final GiftHistoryRepository giftHistoryRepository;
     private final RedisTemplate<String, Object> redisTemplate;
-    private final EmailService emailService;
-    private final TemplateEngine templateEngine;
+    private final ApplicationEventPublisher eventPublisher;
 
     @Override
     public Optional<GiftHistory> payWithPoints(String accountId) {
@@ -38,37 +39,15 @@ public class PaymentServiceImpl implements PaymentService {
         if (!isPayable(sender, priorGiftHistory)) return Optional.empty();
 
         GiftHistory finalGiftHistory = proceedPayment(sender, priorGiftHistory);
-        sendEmailToReceiverAboutGift(finalGiftHistory);
-        return Optional.of(giftHistoryRepository.save(finalGiftHistory));
+        GiftHistory giftHistory = giftHistoryRepository.save(finalGiftHistory);
+        eventPublisher.publishEvent(new GiftHistoryEmailEvent(giftHistory));
+        return Optional.of(giftHistory);
     }
 
     private GiftHistory proceedPayment(Account sender, GiftHistoryRedisHash gift) {
         sender.decreasePoint(gift.getPrice());
         Product product = getProduct(gift.getProductId());
         return gift.saveGiftHistoryWithSender(sender, product);
-    }
-
-    private void sendEmailToReceiverAboutGift(GiftHistory giftHistory) {
-        String message = makeEmailMessageByTemplateEngine(giftHistory);
-        EmailMessage emailMessage = EmailMessage.builder()
-                .to(giftHistory.getReceiverEmail())
-                .subject("선물내역 이메일 수신")
-                .message(message)
-                .build();
-        emailService.send(emailMessage);
-    }
-
-    private String makeEmailMessageByTemplateEngine(GiftHistory giftHistory) {
-        Context context = new Context();
-        context.setVariable("receiverName", giftHistory.getReceiverName());
-        context.setVariable("senderName", giftHistory.getSender().getName());
-        context.setVariable("giftCode", giftHistory.getGiftCode());
-        context.setVariable("expirationDate", giftHistory.getExpirationDate().format(DateTimeFormatter.ofPattern("dd/MM/yyyy")));
-        context.setVariable("brandName", giftHistory.getProduct().getBrandName());
-        context.setVariable("productName", giftHistory.getProduct().getName());
-        context.setVariable("productImage", giftHistory.getProduct().getImage());
-        context.setVariable("message",giftHistory.getMessage());
-        return templateEngine.process("mail/gift-mail-to-receiver", context);
     }
 
     private GiftHistoryRedisHash getGiftHistory(String key) {
